Update eventsubmit.php
This is the webroot component of an inworld LSL script that submits event listings to the OpenSimSearch database.   The LSL component can be found in the corresponding lsl folder.      Be sure to change the server key in both the PHP and LSL scripts to match.      In a netshell it works by dropping a scripted prim on a parcel where the event will take place.   Basically, the assumption is that if you can rez or run the script on a parcel, you are authorized to hold an event there, and it eliminates a lot of complex authentication, group checking, etc, on the server end.